+++ bh-sd-jwt/examples/simple-example.rs
@@ -81,7 +81,7 @@ async fn main() {
     let jwt_payload = IssuerJwt::new(
         "personal_identity_card".to_owned(), // the type of the credential
         iss,
-        holder_signer.public_jwk().unwrap(),
+        Some(holder_signer.public_jwk().unwrap()),
         claims,
     )
     .unwrap();